+++ Source/Aul2AudioFilterPlugin.pas
@@ -29,7 +29,6 @@ uses
   Aul2AudioFilterPluginLimiter,
   Aul2AudioFilterPluginChorus,
   Aul2AudioFilterPluginReverb,
-  Aul2AudioFilterAudioTrace,
   Aul2AudioFilterMonitorBridge;
 
 function GetFilterTable: PFILTER_PLUGIN_TABLE;
@@ -40,7 +39,6 @@ implementation
 
 procedure InitializeFilterPlugin;
 begin
-  AudioTraceInitialize;
   AudioMonitorInitialize;
 end;
 
@@ -69,7 +67,6 @@ begin
       Exit;
     end;
 
-    AudioTraceProcAudio(Audio, SampleNum, ChannelNum);
     AudioMonitorCaptureInput(Audio, SampleNum, ChannelNum);
 
     ProcessDelay(Audio, SampleNum, ChannelNum);
@@ -142,7 +139,6 @@ end;
 procedure FinalizeFilterPlugin;
 begin
   AudioMonitorFinalize;
-  AudioTraceFinalize;
 end;
 
 end.